# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for Predator MQ4 Bots marketplace
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface Bot {
  id: number;
  name: string;
  slug: string;
  description: string;
  longDescription?: string;
  price: number;
  currency: string;
  category: string;
  features: string[];
  imageUrl?: string;
  fileObjectPath?: string | null;
  featured: boolean;
  active: boolean;
  downloadsCount: number;
  createdAt: string;
  updatedAt: string;
}

export interface CreateBotBody {
  name: string;
  slug: string;
  description: string;
  longDescription?: string;
  price: number;
  currency: string;
  category: string;
  features: string[];
  imageUrl?: string;
  fileObjectPath?: string | null;
  featured?: boolean;
  active?: boolean;
}

export type PurchaseStatus =
  (typeof PurchaseStatus)[keyof typeof PurchaseStatus];

export const PurchaseStatus = {
  pending: "pending",
  completed: "completed",
  refunded: "refunded",
} as const;

export interface Purchase {
  id: number;
  userId: string;
  botId: number;
  bot?: Bot;
  status: PurchaseStatus;
  amountPaid: number;
  paymentReference?: string | null;
  downloadCount: number;
  maxDownloads: number;
  secondDownloadExpiresAt?: string | null;
  createdAt: string;
}

export interface CreatePurchaseBody {
  botId: number;
  paymentReference: string;
}

export interface DownloadResponse {
  downloadUrl: string;
  downloadsRemaining: number;
  expiresAt?: string | null;
  expiryNotice?: string;
}

export type MarketplaceStatsCategoriesItem = {
  name: string;
  count: number;
};

export interface MarketplaceStats {
  totalBots: number;
  totalPurchases: number;
  featuredBots: number;
  categories: MarketplaceStatsCategoriesItem[];
}

export interface RequestUploadUrlBody {
  filename: string;
  contentType: string;
}

export interface RequestUploadUrlResponse {
  uploadUrl: string;
  objectPath: string;
}

export type ListBotsParams = {
  category?: string;
  featured?: boolean;
};